adding incomplete software license
Note to 50onRed: Please update this license so that we can use the
mock-jedis code. My company can't legally use unlicensed code. (Search
stackoverflow, etc to see the ramifications of unlicensed code).

Please substitute your own values for <OWNER> and <ORGANIZATION>, and
commit. This is a BSD 3-Clause License from:
http://opensource.org/licenses/BSD-3-Clause

I would greatly appreciate it if you did this, so that we can use the
code. Thanks. :)